SCHED-281: Moved ProgramCalculations to ProgramInfo to avoid circular import.
